pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mdsim"
version = "0.1.0"
description = "Modular MD simulation tool for complex biomolecular systems"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Enes Kelestemur", email = "enes@example.com"},
]
maintainers = [
    {name = "MDSimulation Team", email = "enes@example.com"},
]
keywords = [
    "molecular dynamics", 
    "simulation", 
    "openmm", 
    "protein", 
    "ligand", 
    "drug discovery"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "openmm>=8.1.0",
    "openmmforcefields>=0.12.0",
    "openff-toolkit>=0.15.0",
    "pdbfixer>=1.9",
    "parmed>=4.0.0",
    "mdanalysis>=2.6.0",
    "rdkit>=2023.9.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0.0",
    "networkx>=3.1.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
analysis = [
    "pymol-open-source>=2.5.0",
    "nglview>=3.0.0",
    "py3Dmol>=1.8.0",
]
gpu = [
    "cudatoolkit>=11.0",
]
dev = [
    "pytest>=7.4.0",
]
all = [
    "mdsim[analysis,gpu,dev]"
]

[project.scripts]
mdsim = "mdsim.cli.main:main"

[project.urls]
Homepage = "https://github.com/eneskelestemur/MDSimulation"
Repository = "https://github.com/eneskelestemur/MDSimulation"
"Bug Reports" = "https://github.com/eneskelestemur/MDSimulation/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mdsim*"]
